modified

+++ lib/forecast/model/weather_model.dart
@@ -210,8 +210,8 @@ class Wind {
   Wind({this.speed, this.deg});
 
   Wind.fromJson(Map<String, dynamic> json) {
-    speed = json['speed'];
-    deg = json['deg'];
+    speed = json['speed'].toDouble();
+    deg = json['deg'].toInt();
   }
 
   Map<String, dynamic> toJson() {
